Revert " controllers"
This reverts commit

+++ src/main/java/com/quiz/myquestionsrest/service/QuizService.java
@@ -1,13 +1,16 @@
 package com.quiz.myquestionsrest.service;
 
+import com.quiz.myquestionsrest.exception.EntityNotFoundException;
 import com.quiz.myquestionsrest.model.Quiz;
 import com.quiz.myquestionsrest.repository.QuizRepository;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
 public class QuizService {
@@ -19,18 +22,25 @@ public class QuizService {
     }
 
     public Optional<Quiz> findQuizById(int id) {
-        return Optional.ofNullable(quizRepository.findById(id));
+        return quizRepository.findById(id);
     }
 
     public Optional<Quiz> findQuizByTitle(String title) {
         return quizRepository.findByTitle(title);
     }
 
-    public void saveQuiz(Quiz quiz) {
-        quizRepository.save(quiz);
+    public Quiz saveQuiz(Quiz quiz) throws EntityNotFoundException {
+        if (quiz == null) {
+            throw new EntityNotFoundException("User null");
+        }
+        return quizRepository.save(quiz);
     }
 
     public void deleteById(int id) {
+        if (quizRepository.findById(id).isEmpty()) {
+            log.info("Quiz dos not exist");
+            throw new RuntimeException("Quiz can not a found");
+        }
         quizRepository.deleteById(id);
     }
